feat: restructure create-update generator
