+++ nx-plugin/src/generators/create-update/generator.ts
@@ -6,20 +6,23 @@ import {
   joinPathFragments,
   names,
   readJson,
-  Tree,
-  updateJson,
+  Tree
 } from '@nx/devkit';
 import { execSync } from 'child_process';
-import * as fs from 'fs';
 import * as ora from 'ora';
-import { deepMerge } from '../shared/deepMerge';
 
-import { renderJsonFile } from '../shared/renderJsonFile';
-import { createCreateEndpoint, createUpdateEndpoint } from './endpoint.util';
-import { CreateUpdateGeneratorSchema } from './schema';
-import path = require('path');
+import { GeneratorProcessor } from '../shared/generator.utils';
 import processParams, { GeneratorParameter } from '../shared/parameters.utils';
-import { COMMENT_KEY, OpenAPIUtil } from '../shared/openapi/openapi.utils';
+import { CreateUpdateGeneratorSchema } from './schema';
+import { FeatureModuleStep } from './steps/feature-module.step';
+import { GeneralOpenAPIStep } from './steps/general-openapi.step';
+import { GeneralTranslationsStep } from './steps/general-translations.step';
+import { SearchActionsStep } from './steps/search-actions.step';
+import { SearchComponentStep } from './steps/search-component.step';
+import { SearchEffectsStep } from './steps/search-effects.step';
+import { SearchHTMLStep } from './steps/search-html.step';
+import { SearchTestsStep } from './steps/search-tests.step';
+import path = require('path');
 
 const PARAMETERS: GeneratorParameter<CreateUpdateGeneratorSchema>[] = [
   {
@@ -145,13 +148,22 @@ export async function createUpdateGenerator(
     }
   );
 
-  adaptFeatureModule(tree, options);
+  const generatorProcessor = new GeneratorProcessor();
 
-  addCreateUpdateEventsToSearch(tree, options);
+  generatorProcessor.addStep(new FeatureModuleStep());
+  generatorProcessor.addStep(new GeneralTranslationsStep());
+  generatorProcessor.addStep(new GeneralOpenAPIStep());
 
-  addTranslations(tree, options);
-
-  addFunctionToOpenApi(tree, options);
+  const fileName = names(options.featureName).fileName;
+  const htmlDetailsFilePath = `src/app/${fileName}/pages/${fileName}-search/dialogs/${fileName}-create-update/${fileName}-create-update.component.html`;
+  if (tree.exists(htmlDetailsFilePath)) {
+    generatorProcessor.addStep(new SearchActionsStep());
+    generatorProcessor.addStep(new SearchEffectsStep());
+    generatorProcessor.addStep(new SearchComponentStep());
+    generatorProcessor.addStep(new SearchHTMLStep());
+    generatorProcessor.addStep(new SearchTestsStep());
+  }
+  generatorProcessor.run(tree, options, spinner);
 
   await formatFiles(tree);
 
@@ -177,568 +189,6 @@ export async function createUpdateGenerator(
       stdio: 'inherit',
     });
   };
-}
-
-function addCreateUpdateEventsToSearch(
-  tree: Tree,
-  options: CreateUpdateGeneratorSchema
-) {
-  const fileName = names(options.featureName).fileName;
-  const htmlDetailsFilePath = `src/app/${fileName}/pages/${fileName}-search/dialogs/${fileName}-create-update/${fileName}-create-update.component.html`;
-  if (tree.exists(htmlDetailsFilePath)) {
-    adaptSearchActions(tree, options);
-    adaptSearchEffects(tree, options);
-    adaptSearchComponent(tree, options);
-    adaptSearchHTML(tree, options);
-    adaptSearchTests(tree, options);
-  }
-}
-
-function adaptSearchHTML(tree: Tree, options: CreateUpdateGeneratorSchema) {
-  const fileName = names(options.featureName).fileName;
-  const constantName = names(options.featureName).constantName;
-  const htmlSearchFilePath = `src/app/${fileName}/pages/${fileName}-search/${fileName}-search.component.html`;
-
-  let htmlContent = tree.read(htmlSearchFilePath, 'utf8');
-  htmlContent = htmlContent.replace(
-    '<ocx-interactive-data-view',
-    `<ocx-interactive-data-view 
-      (editItem)="edit($event)"
-      ${options.standalone ? '' : `editPermission="${constantName}#EDIT"`}`
-  );
-  tree.write(htmlSearchFilePath, htmlContent);
-}
-
-function adaptSearchComponent(
-  tree: Tree,
-  options: CreateUpdateGeneratorSchema
-) {
-  const fileName = names(options.featureName).fileName;
-  const className = names(options.featureName).className;
-  const constantName = names(options.featureName).constantName;
-  const filePath = `src/app/${fileName}/pages/${fileName}-search/${fileName}-search.component.ts`;
-
-  let content = tree.read(filePath, 'utf8');
-  content = content.replace(
-    `} from '@onecx/portal-integration-angular';`,
-    `RowListGridData
-    } from '@onecx/portal-integration-angular';`
-  );
-  content = content.replace(
-    'const actions: Action[] = [',
-    `const actions: Action[] = [
-    {
-     labelKey: '${constantName}_CREATE_UPDATE.ACTION.CREATE',
-     icon: PrimeIcons.PLUS,
-     show: 'always',
-     actionCallback: () => this.create(),
-    },`
-  );
-  content = content.replace(
-    'resetSearch',
-    `
-    create() {
-      this.store.dispatch(${className}SearchActions.create${className}ButtonClicked());
-    }
-
-    edit({ id }: RowListGridData) {
-      this.store.dispatch(${className}SearchActions.edit${className}ButtonClicked({ id }));
-    }
-
-    resetSearch`
-  );
-  tree.write(filePath, content);
-}
-
-function adaptSearchActions(tree: Tree, options: CreateUpdateGeneratorSchema) {
-  const fileName = names(options.featureName).fileName;
-  const filePath = `src/app/${fileName}/pages/${fileName}-search/${fileName}-search.actions.ts`;
-  const actionName = names(options.featureName).fileName.replaceAll('-', ' ');
-
-  let content = tree.read(filePath, 'utf8');
-  content = content.replace(
-    'events: {',
-    `events: {
-      'Create ${actionName} button clicked': emptyProps(),
-      'Edit ${actionName} button clicked': props<{
-        id: number | string;
-      }>(),
-      'Create ${actionName} cancelled': emptyProps(),
-      'Update ${actionName} cancelled': emptyProps(),
-      'Create ${actionName} succeeded': emptyProps(),
-      'Update ${actionName} succeeded': emptyProps(),
-      'Create ${actionName} failed': props<{
-        error: string | null;
-      }>(),
-      'Update ${actionName} failed': props<{
-        error: string | null;
-      }>(),      
-    `
-  );
-  tree.write(filePath, content);
-}
-
-function adaptSearchEffects(tree: Tree, options: CreateUpdateGeneratorSchema) {
-  const fileName = names(options.featureName).fileName;
-  const className = names(options.featureName).className;
-  const propertyName = names(options.featureName).propertyName;
-  const constantName = names(options.featureName).constantName;
-  const filePath = `src/app/${fileName}/pages/${fileName}-search/${fileName}-search.effects.ts`;
-
-  let content = tree.read(filePath, 'utf8');
-  content =
-    `import { PortalDialogService } from '@onecx/portal-integration-angular';` +
-    `import { mergeMap } from 'rxjs';` +
-    `import {
-      ${options.dataObjectName},
-      ${options.createRequestName},
-      ${options.updateRequestName},
-    } from 'src/app/shared/generated';` +
-    `import { ${className}CreateUpdateComponent } from './dialogs/${options.featureName}-create-update/${options.featureName}-create-update.component';` +
-    content.replace(
-      'constructor(',
-      `constructor(
-      private portalDialogService: PortalDialogService,`
-    );
-  content = content.replace(
-    'searchByUrl$',
-    `
-      refreshSearch$ = createEffect(() => {
-        return this.actions$.pipe(
-          ofType(
-            ${className}SearchActions.create${className}Succeeded,
-            ${className}SearchActions.update${className}Succeeded
-          ),
-          concatLatestFrom(() => this.store.select(selectSearchCriteria)),
-          switchMap(([, searchCriteria]) => this.performSearch(searchCriteria))
-        );
-    });
-
-    editButtonClicked$ = createEffect(() => {
-    return this.actions$.pipe(
-      ofType(${className}SearchActions.edit${className}ButtonClicked),
-      concatLatestFrom(() =>
-        this.store.select(${propertyName}SearchSelectors.selectResults)
-      ),
-      map(([action, results]) => {
-        return results.find((item) => item.id == action.id);
-      }),
-      mergeMap((itemToEdit) => {
-        return this.portalDialogService.openDialog< ${options.dataObjectName} | undefined>(
-          '${constantName}_CREATE_UPDATE.UPDATE.HEADER',
-          {
-            type: ${className}CreateUpdateComponent,
-            inputs: {
-              vm: {
-                itemToEdit,
-              }
-            },
-          },
-          '${constantName}_CREATE_UPDATE.UPDATE.FORM.SAVE',
-          '${constantName}_CREATE_UPDATE.UPDATE.FORM.CANCEL', {
-            baseZIndex: 100
-          }
-        );
-      }),
-      switchMap((dialogResult) => {
-        if (dialogResult.button == 'secondary') {
-          return of(${className}SearchActions.update${className}Cancelled());
-        }
-        if (!dialogResult?.result) {
-          throw new Error('DialogResult was not set as expected!');
-        }
-        const itemToEditId = dialogResult.result.id;
-        const itemToEdit = {
-            dataObject: dialogResult.result
-        } as ${options.updateRequestName};
-        return this.${propertyName}Service
-          .update${options.dataObjectName}(itemToEditId, itemToEdit)
-          .pipe(
-            map(() => {
-              this.messageService.success({
-                summaryKey: '${constantName}_CREATE_UPDATE.UPDATE.SUCCESS',
-              });
-              return ${className}SearchActions.update${className}Succeeded();
-            })
-          );
-      }),
-      catchError((error) => {
-        this.messageService.error({
-          summaryKey: '${constantName}_CREATE_UPDATE.UPDATE.ERROR',
-        });
-        return of(
-          ${className}SearchActions.update${className}Failed({
-            error,
-          })
-        );
-      })
-    );
-  });
-
-  createButtonClicked$ = createEffect(
-    () => {
-      return this.actions$.pipe(
-        ofType(${className}SearchActions.create${className}ButtonClicked),
-        switchMap(() => {
-          return this.portalDialogService.openDialog< ${options.dataObjectName} | undefined>(
-            '${constantName}_CREATE_UPDATE.CREATE.HEADER',
-            {
-              type: ${className}CreateUpdateComponent,
-              inputs: {
-                vm: {
-                  itemToEdit: {},
-                }
-              },
-            },
-            '${constantName}_CREATE_UPDATE.CREATE.FORM.SAVE',
-            '${constantName}_CREATE_UPDATE.CREATE.FORM.CANCEL', {
-              baseZIndex: 100
-            }
-          );
-        }),
-        switchMap((dialogResult) => {
-          if (dialogResult.button == 'secondary') {
-            return of(${className}SearchActions.create${className}Cancelled());
-          }
-          if (!dialogResult?.result) {
-            throw new Error('DialogResult was not set as expected!');
-          }
-          const toCreateItem = {
-            dataObject: dialogResult.result
-          } as ${options.createRequestName};
-          return this.${propertyName}Service
-            .create${options.dataObjectName}(toCreateItem)
-            .pipe(
-              map(() => {
-                this.messageService.success({
-                  summaryKey: '${constantName}_CREATE_UPDATE.CREATE.SUCCESS',
-                });
-                return ${className}SearchActions.create${className}Succeeded();
-              })
-            );
-        }),
-        catchError((error) => {
-          this.messageService.error({
-            summaryKey: '${constantName}_CREATE_UPDATE.CREATE.ERROR',
-          });
-          return of(
-            ${className}SearchActions.create${className}Failed({
-              error,
-            })
-          );
-        })
-      );
-    }
-  );
-
-    searchByUrl$`
-  );
-  tree.write(filePath, content);
-}
-
-function adaptSearchTests(tree: Tree, options: CreateUpdateGeneratorSchema) {
-  const fileName = names(options.featureName).fileName;
-  const className = names(options.featureName).className;
-  const propertyName = names(options.featureName).propertyName;
-  const filePath = `src/app/${fileName}/pages/${fileName}-search/${fileName}-search.component.spec.ts`;
-
-  let htmlContent = tree.read(filePath, 'utf8');
-
-  htmlContent =
-    `import { PrimeIcons } from 'primeng/api';` +
-    htmlContent.replace(
-      "it('should dispatch export csv data on export action click'",
-      `
-    it('should dispatch edit${className}ButtonClicked action on item edit click', async () => {
-      jest.spyOn(store, 'dispatch');
-  
-      store.overrideSelector(select${className}SearchViewModel, {
-        ...base${className}SearchViewModel,
-        results: [
-          {
-            id: '1',
-            imagePath: '',
-            column_1: 'val_1',
-          },
-        ],
-        columns: [
-          {
-            columnType: ColumnType.STRING,
-            nameKey: 'COLUMN_KEY',
-            id: 'column_1',
-          },
-        ],
-      });
-      store.refreshState();
-  
-      const interactiveDataView =
-        await ${propertyName}Search.getSearchResults();
-      const dataView = await interactiveDataView.getDataView();
-      const dataTable = await dataView.getDataTable();
-      const rowActionButtons = await dataTable.getActionButtons();
-  
-      expect(rowActionButtons.length).toBeGreaterThan(0);
-      let editButton;
-      for (const actionButton of rowActionButtons) {
-        const icon = await actionButton.getAttribute('ng-reflect-icon');
-        expect(icon).toBeTruthy();;
-        if (icon == 'pi pi-pencil') {
-          editButton = actionButton;
-        }
-      }
-      expect(editButton).toBeTruthy();
-      editButton?.click();
-  
-      expect(store.dispatch).toHaveBeenCalledWith(
-        ${className}SearchActions.edit${className}ButtonClicked({ id: '1' })
-      );
-    });
-
-    it('should dispatch create${className}ButtonClicked action on create click', async () => {
-      jest.spyOn(store, 'dispatch');
-
-      const header = await ${propertyName}Search.getHeader();
-      const createButton = await (
-        await header.getPageHeader()
-      ).getInlineActionButtonByIcon(PrimeIcons.PLUS);
-
-      expect(createButton).toBeTruthy();
-      await createButton?.click();
-
-      expect(store.dispatch).toHaveBeenCalledWith(
-        ${className}SearchActions.create${className}ButtonClicked()
-      );
-    });
-
-    it('should export csv data on export action click'`
-    );
-  tree.write(filePath, htmlContent);
-}
-
-function adaptFeatureModule(tree: Tree, options: CreateUpdateGeneratorSchema) {
-  const fileName = names(options.featureName).fileName;
-  const className = names(options.featureName).className;
-  const moduleFilePath = joinPathFragments(
-    'src/app',
-    fileName,
-    fileName + '.module.ts'
-  );
-  let moduleContent = tree.read(moduleFilePath, 'utf8');
-  moduleContent = moduleContent.replace(
-    'declarations: [',
-    `declarations: [${className}CreateUpdateComponent,`
-  );
-  moduleContent = moduleContent.replace(
-    'declarations:',
-    `
-    providers: [providePortalDialogService()],
-    declarations:`
-  );
-  moduleContent = moduleContent.replace(
-    `from '@ngrx/effects';`,
-    `from '@ngrx/effects';  
-     import { ${className}CreateUpdateComponent } from './pages/${fileName}-search/dialogs/${fileName}-create-update/${fileName}-create-update.component';
-     import { providePortalDialogService } from '@onecx/portal-integration-angular';`
-  );
-
-  tree.write(moduleFilePath, moduleContent);
-}
-
-function addTranslations(tree: Tree, options: CreateUpdateGeneratorSchema) {
-  const folderPath = 'src/assets/i18n/';
-  const masterJsonPath = path.resolve(
-    __dirname,
-    './input-files/i18n/master.json.template'
-  );
-
-  const masterJsonContent = renderJsonFile(masterJsonPath, {
-    ...options,
-    featureConstantName: names(options.featureName).constantName,
-    featureClassName: names(options.featureName).className,
-  });
-
-  tree.children(folderPath).forEach((file) => {
-    updateJson(tree, joinPathFragments(folderPath, file), (json) => {
-      const jsonPath = joinPathFragments(
-        path.resolve(__dirname, './input-files/i18n/'),
-        file + '.template'
-      );
-      let jsonContent = {};
-      if (fs.existsSync(jsonPath)) {
-        jsonContent = renderJsonFile(jsonPath, {
-          ...options,
-          featureConstantName: names(options.featureName).constantName,
-          featureClassName: names(options.featureName).className,
-        });
-      }
-
-      json = deepMerge(masterJsonContent, jsonContent, json);
-
-      return json;
-    });
-  });
-}
-
-function addFunctionToOpenApi(
-  tree: Tree,
-  options: CreateUpdateGeneratorSchema
-) {
-  const openApiFolderPath = 'src/assets/swagger';
-  const openApiFiles = tree.children(openApiFolderPath);
-  const bffOpenApiPath = openApiFiles.find((f) => f.endsWith('-bff.yaml'));
-  const bffOpenApiContent = tree.read(
-    joinPathFragments(openApiFolderPath, bffOpenApiPath),
-    'utf8'
-  );
-
-  const dataObjectName = options.dataObjectName;
-  const propertyName = names(options.featureName).propertyName;
-  const createRequestName = options.createRequestName;
-  const createResponseName = options.createResponseName;
-  const updateRequestName = options.updateRequestName;
-  const updateResponseName = options.updateResponseName;
-  const apiServiceName = options.apiServiceName;
-
-  const apiUtil = new OpenAPIUtil(bffOpenApiContent);
-  // Paths
-  apiUtil.paths().set(`/${propertyName}`, {
-    ...createCreateEndpoint(
-      {
-        type: 'post',
-        operationId: `create${dataObjectName}`,
-        tags: [apiServiceName],
-        description: `This operation performs a create.`,
-      },
-      {
-        dataObjectName: dataObjectName,
-        createRequestName: createRequestName,
-        createResponseName: createResponseName,
-      }
-    ),
-  });
-
-  apiUtil.paths().set(
-    `/${propertyName}/{id}`,
-    {
-      ...createUpdateEndpoint(
-        {
-          type: 'put',
-          operationId: `update${dataObjectName}`,
-          tags: [apiServiceName],
-          description: `This operation performs an update.`,
-        },
-        {
-          dataObjectName: dataObjectName,
-          updateRequestSchema: updateRequestName,
-          updateResponseSchema: updateResponseName,
-        }
-      ),
-    },
-    {
-      existStrategy: 'extend',
-    }
-  );
-
-  // Schemas
-  apiUtil
-    .schemas()
-    .set(`${options.createRequestName}`, {
-      type: 'object',
-      properties: {
-        dataObject: {
-          type: 'object',
-          $ref: dataObjectName,
-        },
-        [COMMENT_KEY]: 'ACTION C1: add additional properties here',
-      },
-    })
-    .set(`${options.updateRequestName}`, {
-      type: 'object',
-      properties: {
-        dataObject: {
-          type: 'object',
-          $ref: dataObjectName,
-        },
-        [COMMENT_KEY]: ' ACTION C1: add additional properties here',
-      },
-    })
-    .set(`${options.createResponseName}`, {
-      type: 'object',
-      properties: {
-        id: {
-          type: 'string',
-        },
-        changeMe: {
-          type: 'string',
-        },
-        [COMMENT_KEY]: 'ACTION C1: add additional properties here',
-      },
-    })
-    .set(`${options.updateResponseName}`, {
-      type: 'object',
-      properties: {
-        id: {
-          type: 'string',
-        },
-        changeMe: {
-          type: 'string',
-        },
-        [COMMENT_KEY]: ' ACTION C1: add additional properties here',
-      },
-    });
-
-  apiUtil.schemas().set('ProblemDetailResponse', {
-    type: 'object',
-    properties: {
-      errorCode: {
-        type: 'string',
-      },
-      detail: {
-        type: 'string',
-      },
-      params: {
-        type: 'array',
-        items: {
-          $ref: '#/components/schemas/ProblemDetailParam',
-        },
-      },
-      invalidParams: {
-        type: 'array',
-        items: {
-          $ref: '#/components/schemas/ProblemDetailInvalidParam',
-        },
-      },
-    },
-  });
-
-  apiUtil.schemas().set('ProblemDetailParam', {
-    type: 'object',
-    properties: {
-      key: {
-        type: 'string',
-      },
-      value: {
-        type: 'string',
-      },
-    },
-  });
-
-  apiUtil.schemas().set('ProblemDetailInvalidParam', {
-    type: 'object',
-    properties: {
-      name: {
-        type: 'string',
-      },
-      message: {
-        type: 'string',
-      },
-    },
-  });
-
-  tree.write(
-    joinPathFragments(openApiFolderPath, bffOpenApiPath),
-    apiUtil.finalize()
-  );
 }
 
 export default createUpdateGenerator;
